Create scripts to handle creation of VPN
Applications like MATLAB require a VPN to connect to a license server on
campus. I want to use matlab on my home machine while connected to it
through ssh. This means I need a more advanced script to set up the vpn
only for the application. this will allow me to continue using ssh
uninterrupted.